// frontend/src/types/exam.ts
export type AdminTransaction = {
  id: string;
  code: string;
  type: 'MEMBERSHIP' | 'ADDON';
  amount: number;
  method: string;
  status: 'PENDING' | 'PAID' | 'REJECTED';
  createdAt: string;
  package: { name: string; category: string };
  addon?: { name: string } | null;
  user: { name: string; email: string };
  proofUrl?: string | null;
  description?: string | null;
  activatedAt?: string | null;
  expiresAt?: string | null;
};